refactor[OO]: moving the server logic to a new server structure
- avoids multiple use of cfg in main
- server logic is no longer scattered between main and server
